to next frontent

+++ app/Services/Library/Actions/DeleteFile.php
@@ -12,37 +12,40 @@ use Throwable;
 
 final class DeleteFile implements LibraryDeleteContract
 {
-    use Destroyable;
+  use Destroyable;
 
-    /**
-     * Destroy File
-     */
-    public function handle(Library $library): bool
-    {
-        if ($library->exists) {
-            $file = LibDeleteData::fromArray([
-                'file_url' => $library->image_url,
-                'file_conversions' => [
-                    'lg' => $library->conversions['lg_path'],
-                    'md' => $library->conversions['md_path'],
-                    'sm' => $library->conversions['sm_path'],
-                ],
-            ]);
-            try {
-                $this->destroy(
-                    file: $file,
-                );
-                session()->now('success', 'Destroyed File From Storage');
-            } catch (Throwable $e) {
-                report($e);
-                session()->now('warning', 'Could not delete file');
-            }
-            $library->delete();
-            cache()->store('library')->clear();
+  /**
+   * Destroy File
+   */
+  public function handle(Library $library, bool $destroy = true): bool
+  {
+    if ($library->exists) {
+      $file = LibDeleteData::fromArray([
+        'file_url' => $library->image_url,
+        'file_conversions' => [
+          'lg' => $library->conversions['lg_path'] ?? $library->conversions['lg'],
+          'md' => $library->conversions['md_path'] ?? $library->conversions['md'],
+          'sm' => $library->conversions['sm_path'] ?? $library->conversions['sm'],
+        ],
+      ]);
+      try {
+        $this->destroy(
+          file: $file,
+        );
+        session(['success', 'Destroyed File From Storage']);
+      } catch (Throwable $e) {
+        report($e);
+        session(['warning', 'Could not delete file']);
+      }
+      if (true === $destroy) {
+        $library->delete();
+      }
 
-            return true;
-        }
+      cache()->flush();
 
-        return false;
+      return true;
     }
+
+    return false;
+  }
 }
